Isolated weather icon, isolated helper functions for Getting my current position, and fetching weather.
