+++ src/Helpers/IsNight.js
@@ -1,0 +1,9 @@
+// Checks to see if its currently night.
+
+export const IsNight = () => {
+  var date = new Date();
+
+  return (
+    date.getHours() > 22 || date.getHours() < 6
+  );
+}
